Implemented auto-adding of new pages to navbar and routing.
The official documentation for Webpack's require.context() is
[here](https://webpack.js.org/guides/dependency-management/#requirecontext)
but is incredibly lacking, to say the least. I assume it might be
sufficient for someone who has read enough of the documentation up to
this point, and even then I'm not a huge fan of the design, as it
appears-- why is keys called by `.keys()` but resolve can just be called
by using the thing precisely as it is returned, as opposed to something
like `.resolve()`?

Anyways, not immmensely proud or ashamed of this but basically all the
code I ended up with was inspired by slightly different (and incorrect)
versions that the google search AI spit out when I was trying to figure
out what the types and stuff surrounding require.context were.

Definitely got a nudge in the right direction from
[this stackoverflow question]
(https://stackoverflow.com/questions/29875869/react-jsx-dynamic-component-name),
which also helps to explain some of the complexity in the allPages function,
since the routing needs a custom component, so I somehow needed to
dynamically return all the imported page components as functions instead
of just strings.